Adding detail.html page for the item management
As per the feedback and suggestion, I have added a details.html page, which includes the person's detail. It navigates to other page which has the same view as wiki news. As the data base is not connected, for now, I have created a template format of how data flows in the details.html page.